# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CFR Content API",
        "description": "Public API for retrieving website content",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1"
        }
    ],
    "paths": {
...
        "/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "List all published activities",
                "description": "Returns a paginated list of published experience activities. Supports filtering by category name.",
                "operationId": "fa04cbb03041bb875d6212b89bae56b0",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of activities"
                    }
                }
            }
        },
        "/activities/{slug}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get specific activity details",
                "description": "Returns all the details of a single published experience activity by its slug.",
                "operationId": "ac3d8bdf948a693313295e8372e47104",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the activity",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity details"
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                }
            }
        },
...
        "/pages/{slug}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get page with all sections",
                "description": "Returns a complete page with hero section and all dynamic sections with their content filtered by display_fields",
                "operationId": "01cf406b8dbd1ebf7e316d5cb6877490",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Page slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page with sections"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}